fix: linting and formatting

+++ src/nvim/os/shell.c
@@ -1245,7 +1245,7 @@ static size_t write_output(char *output, size_t remaining, bool eof)
         ml_append(curwin->w_cursor.lnum++, new_line, (int)strlen(new_line) + 1, false);
       } else if (curbuf->b_op_start.lnum == cur_ln_nr) {
         char *old_line = ml_get(cur_ln_nr);
-        int old_chars_indx = (curbuf->b_op_start.col);
+        size_t old_chars_indx = (size_t)curbuf->b_op_start.col;
         char *old_txt = xcalloc(old_chars_indx, sizeof(char));
         memcpy(old_txt, old_line, old_chars_indx * sizeof(char));
         char *new_line = concat_str(old_txt, output);
